requirements.txt
markdownify==0.6.1
markdown==3.3.4
discord.py==1.6.0
praw==7.1.0
opml==0.5
feedparser==6.0.2
python-dotenv==0.19.1
python-telegram-bot==20.5
